Renamed scanDownloaded to removeExpired and refactored. #34
